add polars elementwise check support
